+++ __tests__/__system__/cli/issue/cli.issue.command-with-options.system.test.ts
@@ -9,12 +9,11 @@
 *
 */
 
-import {runCliScript} from "../../../__src__/TestUtils";
-import {ITestEnvironment} from "../../../__src__/environment/doc/response/ITestEnvironment";
-import {TestEnvironment} from "../../../__src__/environment/TestEnvironment";
+import { ITestEnvironment, TestEnvironment, runCliScript } from "@zowe/cli-test-utils";
+import { ITestPropertiesSchema } from "../../../__src__/environment/doc/ITestPropertiesSchema";
 
 // Test environment will be populated in the "beforeAll"
-let TEST_ENVIRONMENT: ITestEnvironment;
+let TEST_ENVIRONMENT: ITestEnvironment<ITestPropertiesSchema>;
 
 describe("zowe-cli-sample issue command-with-options", () => {
 
@@ -29,8 +28,8 @@ describe("zowe-cli-sample issue command-with-options", () => {
         await TestEnvironment.cleanUp(TEST_ENVIRONMENT);
     });
 
-    it("should issue a response with the number and string inputted", async () => {
-        const response = await runCliScript(__dirname + "/__scripts__/issue_command_with_options.sh", TEST_ENVIRONMENT,
+    it("should issue a response with the number and string inputted", () => {
+        const response = runCliScript(__dirname + "/__scripts__/issue_command_with_options.sh", TEST_ENVIRONMENT,
             [1, "\"teststring!\""]);
 
         expect(response.stderr.toString()).toBe("");